+++ django_boost/__init__.py
@@ -1,5 +1,10 @@
-from django.utils.version import get_version
+VERSION = (1, 2, 3)
 
-VERSION = (1, 2, 0, 'final', 0)
+def get_version(version=VERSION):
+    if version[2] == 0:
+        return "%s.%s" % version[:2]
+    return "%s.%s.%s" % version
 
-__version__ = get_version(VERSION)
+__version__ = get_version()
+
+default_app_config = 'django_boost.apps.DjangoBoostConfig'