+++ django_boost/urls/__init__.py
@@ -4,11 +4,13 @@ from django_boost.urls.converters import (
     HexIntConverter, HexStrConverter,
     OctIntConverter, OctStrConverter,
     register_boost_converters)
+from django_boost.urls.static import include_static_files
 
 
 __all__ = ['UrlSet', 'HexIntConverter', 'HexStrConverter',
            'OctIntConverter', 'OctStrConverter', 'BinIntConverter',
-           'BinStrConverter', 'DateConverter', 'register_boost_converters']
+           'BinStrConverter', 'DateConverter', 'register_boost_converters',
+           'include_static_files']
 
 
 class UrlSet: