+++ django_boost/views/generic.py
@@ -2,12 +2,12 @@ from django.urls import path, reverse
 
 from django_boost.views.base import (
     CreateView, DeleteView, DetailView, FormView,
-    ListView, TemplateView, UpdateView, View)
+    ListView, StaticView, TemplateView, UpdateView, View)
 from django_boost.views.mixins import JsonRequestMixin, JsonResponseMixin
 
 __all__ = ["CreateView", "DeleteView", "DetailView",
            "FormView", "JsonView", "ListView", "TemplateView",
-           "UpdateView", "View"]
+           "UpdateView", "View", "StaticView"]
 __views__ = ["BaseModelCLUDViews", "ModelCRUDViews"]
 
 __all__ += __views__