+++ tests/tests/urls_converters/tests.py
@@ -23,7 +23,7 @@ ROOT_PATH = os.path.dirname(__file__)
 )
 class TestConverter(TestCase):
 
-    def test_converters(self):
+    def test_path_converters(self):
         case = [('bin', '1010'),
                 ('bin', 12),
                 ('oct', '7'),
@@ -33,11 +33,25 @@ class TestConverter(TestCase):
                 ('bin_str', '1010'),
                 ('oct_str', '236'),
                 ('hex_str', '234'),
-                ('date', '2020/2/29')]
+                ('float', 1.1),
+                ('float', '1.1'),
+                ('float', 1),
+                ('float', '1'),
+                ('date', '2020/2/29'),
+                ]
         for name, value in case:
             url = reverse(name, kwargs={name: value})
             response = self.client.get(url)
             self.assertStatusCodeEqual(response, 200)
+
+    def test_failed_case(self):
+        from django.urls.exceptions import NoReverseMatch
+
+        with self.assertRaises(NoReverseMatch):
+            reverse('float', kwargs={'float': '1.'})
+
+        with self.assertRaises(NoReverseMatch):
+            reverse('date', kwargs={'date': '2019/2/29'})
 
 
 class TestRegex(TestCase):